# microservices/api-gateway/openapi.yml
openapi: 3.0.3
info:
  title: POS Multi-Magasins - API Gateway
  description: |
    API Gateway pour le système POS Multi-Magasins avec architecture microservices.
    
    ## Authentification
    
    ### API Keys (Kong Gateway)
    - **Header**: `X-API-KEY`
    - **Valeur**: `pos-microservices-2025`
    
    ### JWT Authentication (Services clients)
    - **Header**: `Authorization`
    - **Format**: `Bearer <jwt_token>`
    - **Endpoint de login**: `POST /api/v1/customers/auth/login`
    
    ## Architecture Microservices
    
    Le système est composé de 7 microservices :
    
    ### Services Métier (extraits du monolithe)
    1. **Product Service** (Port 8001) - Catalogue produits
    2. **Inventory Service** (Port 8002) - Gestion des stocks  
    3. **Sales Service** (Port 8003) - Transactions POS
    4. **Reporting Service** (Port 8004) - Rapports analytiques
    
    ### Services E-commerce (nouveaux)
    5. **Customer Service** (Port 8005) - Comptes clients
    6. **Cart Service** (Port 8006) - Panier d'achats
    7. **Order Service** (Port 8007) - Gestion des commandes
    
    ## Load Balancing
    
    Le Cart Service utilise un load balancer avec 3 instances (algorithme least-connections).
    
    ## Monitoring
    
    - **Prometheus**: Métriques sur `/metrics`
    - **Grafana**: Dashboard Kong Gateway
    - **Health checks**: `/health` sur chaque service
    
  version: "1.0.0"
  contact:
    name: LOG430 Lab 5 Team
    email: lab430@example.com
  license:
...
servers:
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: Clé API pour accéder aux microservices via Kong Gateway
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT pour l'authentification des clients
  schemas:
...
    Customer:
      type: object
      required: [email, nom, prenom]
      properties:
        id:
          type: integer
          description: ID unique du client
        email:
          type: string
          format: email
          description: Email du client
        nom:
          type: string
          description: Nom du client
        prenom:
          type: string
          description: Prénom du client
        telephone:
          type: string
          description: Numéro de téléphone
        date_naissance:
          type: string
          format: date
          description: Date de naissance
        actif:
          type: boolean
          description: Compte actif
    
    # Modèles Cart Service
...
paths:
...
  /api/v1/customers/auth/register:
    post:
      tags: [Authentication]
      summary: Créer un compte client
      description: Enregistrer un nouveau client dans le système
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, nom, prenom]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                nom:
                  type: string
                prenom:
                  type: string
                telephone:
                  type: string
                date_naissance:
                  type: string
                  format: date
      responses:
        '201':
          description: Compte créé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  customer:
                    $ref: '#/components/schemas/Customer'
  /api/v1/customers/auth/login:
    post:
      tags: [Authentication]
      summary: Authentifier un client
      description: Connecter un client et obtenir un token JWT
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Authentification réussie
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Token JWT pour les requêtes authentifiées
                  customer:
                    $ref: '#/components/schemas/Customer'
        '401':
          description: Identifiants invalides

  # Cart Service Routes (Load Balanced)
...
  /api/v1/customers/{customer_id}:
    get:
      tags: [Customers]
      summary: Récupérer un client
      description: Obtenir les informations d'un client
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Informations du client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'

  # Cart Service Routes (Load Balanced)
...